requirements.txt
django
djangorestframework
pandas
binance
bta-lib
requests